+++ firmware/tsconfig.json
@@ -200,7 +200,9 @@
 		"/workspaces/stack-chan/firmware/extern/avatar/src/marquee-label.ts",
 		"/workspaces/stack-chan/firmware/extern/rs30x/src/rs30x.ts",
 		"/workspaces/stack-chan/firmware/stackchan/main.ts",
+		"/workspaces/stack-chan/firmware/stackchan/stackchan-mod.ts",
 		"/workspaces/stack-chan/firmware/stackchan/robot.ts",
+		"/workspaces/stack-chan/firmware/stackchan/button.ts",
 		"/workspaces/stack-chan/firmware/stackchan/rs30x-driver.ts",
 		"/workspaces/stack-chan/firmware/stackchan/sg90-driver.ts"
 	]
